Remove base pipeline, specify hyperparameters passing
